Template folder cleanup (v3)
- Structure the template files in the same way as the scaffolded files
- Provide a more descriptive name and comment for each template
- Fix references to interfaces being implemented
