+++ pkg/plugin/v3/scaffolds/internal/templates/config/manager/config.go
@@ -24,7 +24,7 @@ import (
 
 var _ file.Template = &Config{}
 
-// Config scaffolds yaml config for the manager.
+// Config scaffolds a file that defines the namespace and the manager deployment
 type Config struct {
 	file.TemplateMixin
 
@@ -32,7 +32,7 @@ type Config struct {
 	Image string
 }
 
-// SetTemplateDefaults implements input.Template
+// SetTemplateDefaults implements file.Template
 func (f *Config) SetTemplateDefaults() error {
 	if f.Path == "" {
 		f.Path = filepath.Join("config", "manager", "manager.yaml")
