+++ pkg/plugin/v3/scaffolds/internal/templates/config/rbac/auth_proxy_role_binding.go
@@ -22,29 +22,34 @@ import (
 	"sigs.k8s.io/kubebuilder/pkg/model/file"
 )
 
-var _ file.Template = &ClientClusterRole{}
+var _ file.Template = &AuthProxyRoleBinding{}
 
-// ClientClusterRole scaffolds the config/rbac/client_clusterrole.yaml file
-type ClientClusterRole struct {
+// AuthProxyRoleBinding scaffolds a file that defines the role binding for the auth proxy
+type AuthProxyRoleBinding struct {
 	file.TemplateMixin
 }
 
-// SetTemplateDefaults implements input.Template
-func (f *ClientClusterRole) SetTemplateDefaults() error {
+// SetTemplateDefaults implements file.Template
+func (f *AuthProxyRoleBinding) SetTemplateDefaults() error {
 	if f.Path == "" {
-		f.Path = filepath.Join("config", "rbac", "auth_proxy_client_clusterrole.yaml")
+		f.Path = filepath.Join("config", "rbac", "auth_proxy_role_binding.yaml")
 	}
 
-	f.TemplateBody = clientClusterRoleTemplate
+	f.TemplateBody = proxyRoleBindinggTemplate
 
 	return nil
 }
 
-const clientClusterRoleTemplate = `apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
+const proxyRoleBindinggTemplate = `apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRoleBinding
 metadata:
-  name: metrics-reader
-rules:
-- nonResourceURLs: ["/metrics"]
-  verbs: ["get"]
+  name: proxy-rolebinding
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: ClusterRole
+  name: proxy-role
+subjects:
+- kind: ServiceAccount
+  name: default
+  namespace: system
 `
