+++ pkg/plugin/v3/scaffolds/internal/templates/config/rbac/leader_election_role.go
@@ -24,12 +24,12 @@ import (
 
 var _ file.Template = &LeaderElectionRole{}
 
-// LeaderElectionRole scaffolds the config/rbac/leader_election_role.yaml file
+// LeaderElectionRole scaffolds a file that defines the role that allows leader election
 type LeaderElectionRole struct {
 	file.TemplateMixin
 }
 
-// SetTemplateDefaults implements input.Template
+// SetTemplateDefaults implements file.Template
 func (f *LeaderElectionRole) SetTemplateDefaults() error {
 	if f.Path == "" {
 		f.Path = filepath.Join("config", "rbac", "leader_election_role.yaml")
